move dht and remove new p2p message
