+++ node/src/bin.rs
@@ -10,6 +10,7 @@ use tokio::task::JoinHandle;
 mod blocks;
 mod cli;
 mod client;
+mod dht;
 mod id;
 mod message;
 mod net;
@@ -17,7 +18,6 @@ mod node;
 mod server;
 mod stats;
 mod utils;
-mod dht;
 
 /// Turn the tokio-console subscriber on or off by passing `true` or `false`
 #[macro_export]
